nixd/Server: sync the draft store
