+++ nixd/include/nixd/Server/Role.h
@@ -1,0 +1,13 @@
+#pragma once
+
+namespace nixd {
+
+enum class ServerRole {
+  /// Parent process of the server
+  Controller,
+  /// Child process
+  Evaluator,
+  OptionProvider,
+};
+
+} // namespace nixd
